+++ client/src/API/user.service.ts
@@ -1,4 +1,12 @@
+import React, { useContext } from "react";
 import axios from "axios";
+import { AuthContext } from "../context/AuthContext";
+
+const SERVER_HOST_URL =
+  process.env.NODE_ENV === "production"
+    ? "https://sub-u.onrender.com"
+    : "http://localhost:8000";
+
 export async function userReg(
   username: string,
   email: string,
@@ -7,7 +15,7 @@ export async function userReg(
   let data: resData = {};
   try {
     data = await axios
-      .post("https://sub-u.onrender.com/api/users/reg", {
+      .post(`${SERVER_HOST_URL}/api/users/reg`, {
         email,
         password,
         name: username,
@@ -21,18 +29,29 @@ export async function userLogin(email: string, password: string) {
   let data: resData = {};
   try {
     data = await axios
-      .post("https://sub-u.onrender.com/api/users/login", { email, password })
+      .post(`${SERVER_HOST_URL}/api/users/login`, { email, password })
       .then((res) => res?.data);
   } finally {
     return data;
   }
 }
+
+export async function checkToken(userToken: string) {
+  let data: resData = {};
+  try {
+    data = await axios
+      .get(`${SERVER_HOST_URL}/api/users/auth`, {
+        headers: { Authorization: "test " + userToken },
+      })
+      .then((res) => res?.data);
+  } finally {
+    return data;
+  }
+}
+
 interface resData {
   token?: string;
   message?: string;
 }
 
-export default {
-  userLogin,
-  userReg,
-};
+export default { checkToken, userLogin, userReg };