+++ client/src/components/Header.tsx
@@ -1,11 +1,14 @@
-import React, { HTMLAttributes } from "react";
+import React, { HTMLAttributes, useContext } from "react";
 import "../styles/Header.css";
 import { MenuItemI } from "../types";
 import { Link } from "react-router-dom";
+import { AuthContext } from "../context/AuthContext";
 
 function Header(
   props: HTMLAttributes<HTMLDivElement> & { menuList: MenuItemI[] }
 ) {
+  const { isAuth } = useContext(AuthContext);
+
   return (
     <div className="Header">
       <Link to="/" className="logo">
@@ -19,6 +22,16 @@ function Header(
             </li>
           );
         })}
+        {isAuth && (
+          <li>
+            <img
+              src="https://cdn.onlinewebfonts.com/svg/img_568656.png"
+              alt="ava"
+              height={72}
+              width={72}
+            />
+          </li>
+        )}
       </ul>
     </div>
   );